% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymsam-package.R
\docType{package}
\name{asymsam-package}
\alias{asymsam}
\alias{asymsam-package}
\title{asymsam: What the Package Does (One Line, Title Case)}
\description{
Research compendium for a workng paper.
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
