% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_locations.R
\name{data_path}
\alias{data_path}
\alias{AAO}
\alias{SST}
\alias{ERA5}
\alias{OLR}
\alias{CMAP}
\title{Data locations}
\usage{
data_path(type = c("raw", "derived"), ...)

AAO()

SST()

ERA5()

OLR()

CMAP()
}
\arguments{
\item{type}{character indicating the type of data. The resulting base will be type_data}

\item{...}{characters to be contactenadted with \code{\link[=path.file]{path.file()}}}
}
\description{
Functions that return the location of different datasets. They check that
relevant file exist and then return it with a class that has a neat print method.
}
