% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-tags.R
\name{facet_tags}
\alias{facet_tags}
\title{Adds tags to facets}
\usage{
facet_tags(
  position = "tl",
  tag_pool = letters,
  open = "",
  close = ")",
  stroke = 0,
  stroke.color = "white",
  color = "black"
)
}
\description{
Adds tags to facets
}
