% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-management.R
\name{nc_append}
\alias{nc_append}
\alias{nc_concatenate}
\title{Tools to manage netcdf4 files}
\usage{
nc_append(in_files, out_file, verbose = TRUE)

nc_concatenate(in_files, out_file, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{in_files}{character vector of files}

\item{out_file}{file to save output}

\item{verbose}{print out commands}

\item{overwrite}{wether to overwrite exiting files}
}
\description{
Wrappers around nco commands .
}
