% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-tags.R
\name{tag_facets}
\alias{tag_facets}
\title{Adds tags to facets}
\usage{
tag_facets(
  position = "tl",
  tag_levels = "a",
  tag_pool = NULL,
  tag_prefix = "",
  tag_suffix = ")",
  stroke = 0.15,
  stroke.color = "white",
  color = "black"
)
}
\description{
Adds tags to facets
}
