% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_asymsam}
\alias{theme_asymsam}
\title{Main theme used in the paper}
\usage{
theme_asymsam(base_size = 16)
}
\arguments{
\item{base_size}{base font size, given in pts}
}
\description{
Based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} but with minimal (wink) modifications.
}
